pyproject.toml
[tool.poetry]
name = "shamir-image-shares"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11, <3.12"
mnemonic = "^0.20"
tabulate = "^0.9.0"
click = "^8.1.7"
bip-utils = "^2.8.0"
pytest = "^7.4.3"
pillow = "^10.1.0"
requests = "^2.31.0"
numpy = "^1.26.2"
attrs = "^23.1.0"
